Modify data type of matrix from float to int

+++ matrix.h
@@ -5,7 +5,7 @@
 
 /* predefined shortcut */
 #define DECLARE_MATRIX(col, row) \
-    typedef struct { float values[col][row]; } Mat ## col ## x ## row
+    typedef struct { int values[col][row]; } Mat ## col ## x ## row
 DECLARE_MATRIX(3, 3);
 DECLARE_MATRIX(4, 4);
